[FEATURE]#122 : 현재 시간 체크 로직 스크린 내부로 수정
